chore: moved auth helpers in its own module
