+++ src/utils/helpers/auth/isFieldRequired.ts
